matching1 adopted for Linux

+++ db/MySQL/MySQLField.cpp
@@ -1,4 +1,5 @@
 #include <vector>
+#include <cstdint>
 #include <basic/TextConv.h>
 #include "MySQLField.h"
 #include "MySQLStmtData.h"
@@ -61,7 +62,7 @@ std::wstring CMySQLField::getFieldNameW() const {
 	std::vector<wchar_t> field_name_w;
 	MYSQL_FIELD *mysql_field = this->getMySQLFieldHandle();
 
-	UTF8_ToUCS16(mysql_field->org_name, mysql_field->org_name_length, field_name_w);
+	UTF8_ToUCS16(mysql_field->org_name, (int)mysql_field->org_name_length, field_name_w);
 	field_name_w.push_back(0);
 	
 	return std::wstring(&field_name_w[0]);
@@ -78,7 +79,7 @@ std::wstring CMySQLField::getFieldAliasW() const {
 	std::vector<wchar_t> field_name_w;
 	MYSQL_FIELD *mysql_field = this->getMySQLFieldHandle();
 
-	UTF8_ToUCS16(mysql_field->name, mysql_field->name_length, field_name_w);
+	UTF8_ToUCS16(mysql_field->name, (int)mysql_field->name_length, field_name_w);
 	field_name_w.push_back(0);
 
 	return std::wstring(&field_name_w[0]);
@@ -100,7 +101,7 @@ std::wstring CMySQLField::getTableNameW() const {
 	std::vector<wchar_t> table_name_w;
 	MYSQL_FIELD *mysql_field = this->getMySQLFieldHandle();
 
-	UTF8_ToUCS16(mysql_field->org_table, mysql_field->org_table_length, table_name_w);
+	UTF8_ToUCS16(mysql_field->org_table, (int)mysql_field->org_table_length, table_name_w);
 	table_name_w.push_back(0);
 
 	return std::wstring(&table_name_w[0]);
